--- API/QAQ_api/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from . import views
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('', views.PostList.as_view()),
-    path('<int:pk>/', views.PostDetail.as_view()),]
