+++ api/admin.py
@@ -1,4 +1,5 @@
 from django.contrib import admin
-from . models import Post
+from api.models import Post
+
 admin.site.register(Post)
 # Register your models here.
