+++ docker/docker-compose.yml
@@ -23,6 +23,7 @@ services:
     volumes:
       - ../client:/var/www/apache/client:ro
       - ../server:/var/www/apache/server:ro
+      - ../api:/var/www/apache/api:ro
     entrypoint: [ "bash", "scripts/run-local.bash" ]
 volumes:
   sweng-db:
