+++ scripts/run-migrations.sh
@@ -1,0 +1,39 @@
+#!/bin/bash
+#
+# This script exists as migrations cannot be performed within the
+# docker container as the server, client and api folders are mounted as
+# read only to prevent messed up permissions. To ameliorate this, this
+# script temporarily starts the database, and runs a specified script,
+# which is 'migrate' by default.
+#
+# Example Usage
+# =============
+#
+# $ sh scripts/run-migrations.sh
+#   > No changes detected
+#   > Operations to perform:
+#   >   Apply all migrations: admin, api, auth, contenttypes, sessions
+#   > Running migrations:
+#   >   No migrations to apply.
+#
+# $ sh scripts/run-migrations.sh sqlmigrate api 0001
+#   > BEGIN;
+#   > --
+#   > -- Create model Post
+#   > --
+#   > CREATE TABLE "api_post" ("id" bigint NOT NULL PRIMARY KEY GENERATED BY ... etc
+#   > COMMIT;
+
+. scripts/install.sh
+cd docker; docker compose up -d database; cd ../
+container=$(docker container list | awk '/database/ { print $1 }')
+host=$(docker inspect -f '{{ range .NetworkSettings.Networks }}{{ .IPAddress }}{{end}}' $container)
+export POSTGRES_HOST=$host
+(if [ $# = 0 ]
+ then
+     python3 manage.py makemigrations || exit 1
+     python3 manage.py migrate
+ else
+     python3 manage.py "$@"
+ fi)
+docker compose down database
