+++ server/admin.py
@@ -1,0 +1,2 @@
+
+# Register your models here.
