+++ server/urls.py
@@ -1,23 +1,6 @@
-"""QAQ_project URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-from django.contrib import admin
 from django.urls import path
-from django.contrib import admin
-from django.urls import include, path
+from api import views
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('api/', include('QAQ_api.urls')),]
+    path('', views.PostList.as_view()),
+    path('<int:pk>/', views.PostDetail.as_view()),]
